+++ addOns/reports/src/test/java/org/zaproxy/addon/reports/ExtensionReportsUnitTest.java
@@ -37,6 +37,7 @@ import java.util.Arrays;
 import java.util.Date;
 import java.util.List;
 import java.util.Locale;
+import java.util.Map;
 import javax.xml.XMLConstants;
 import javax.xml.parsers.DocumentBuilder;
 import javax.xml.parsers.DocumentBuilderFactory;
@@ -93,6 +94,93 @@ class ExtensionReportsUnitTest {
         Model.getSingleton().getOptionsParam().load(new ZapXmlConfiguration());
 
         Constant.PROGRAM_VERSION = "Dev Build";
+    }
+
+    @Test
+    void shouldReturnDefaultNoCounts() {
+        // Given
+        ExtensionReports extRep = new ExtensionReports();
+        AlertNode root = new AlertNode(0, "Alerts");
+
+        // When
+        Map<Integer, Integer> counts = extRep.getAlertCountsByRule(root);
+
+        // Then
+        assertThat(counts.size(), is(equalTo(0)));
+    }
+
+    @Test
+    void shouldReturnExpectedCounts() {
+        // Given
+        ExtensionReports extRep = new ExtensionReports();
+        AlertNode root = new AlertNode(0, "Alerts");
+        root.add(newAlertNode(1, Alert.RISK_HIGH, "Alert High 1", "https://www.example.com", 1));
+        root.add(
+                newAlertNode(2, Alert.RISK_MEDIUM, "Alert Medium 1", "https://www.example.com", 2));
+        root.add(newAlertNode(3, Alert.RISK_LOW, "Alert Low 1", "https://www.example.com", 4));
+        root.add(newAlertNode(4, Alert.RISK_LOW, "Alert Low 2", "https://www.example.com", 8));
+
+        // When
+        Map<Integer, Integer> counts = extRep.getAlertCountsByRule(root);
+
+        // Then
+        assertThat(counts.size(), is(equalTo(4)));
+        assertThat(counts.get(1), is(equalTo(1)));
+        assertThat(counts.get(2), is(equalTo(2)));
+        assertThat(counts.get(3), is(equalTo(4)));
+        assertThat(counts.get(4), is(equalTo(8)));
+    }
+
+    @Test
+    void shouldReturnExpectedCountsWithSameAlertWithDifferentRisk() {
+        // Given
+        ExtensionReports extRep = new ExtensionReports();
+        AlertNode root = new AlertNode(0, "Alerts");
+        root.add(newAlertNode(1, Alert.RISK_HIGH, "Alert High 1", "https://www.example.com", 1));
+        root.add(
+                newAlertNode(1, Alert.RISK_MEDIUM, "Alert Medium 1", "https://www.example.com", 2));
+        root.add(newAlertNode(2, Alert.RISK_LOW, "Alert Low 1", "https://www.example.com", 4));
+        root.add(newAlertNode(2, Alert.RISK_LOW, "Alert Low 2", "https://www.example.com", 8));
+
+        // When
+        Map<Integer, Integer> counts = extRep.getAlertCountsByRule(root);
+
+        // Then
+        assertThat(counts.size(), is(equalTo(2)));
+        assertThat(counts.get(1), is(equalTo(3)));
+        assertThat(counts.get(2), is(equalTo(12)));
+    }
+
+    @Test
+    void shouldReturnExpectedCountsIgnoringFalsePositives() {
+        // Given
+        ExtensionReports extRep = new ExtensionReports();
+        AlertNode root = new AlertNode(0, "Alerts");
+        root.add(newAlertNode(1, Alert.RISK_HIGH, "Alert High 1", "https://www.example.com", 1));
+        root.add(newAlertNode(1, -1, "Alert Medium 1", "https://www.example.com", 2));
+        root.add(newAlertNode(2, Alert.RISK_LOW, "Alert Low 1", "https://www.example.com", 4));
+        root.add(newAlertNode(2, -1, "Alert Low 2", "https://www.example.com", 8));
+        root.add(newAlertNode(3, -1, "Alert Low 3", "https://www.example.com", 16));
+
+        // When
+        Map<Integer, Integer> counts = extRep.getAlertCountsByRule(root);
+
+        // Then
+        assertThat(counts.size(), is(equalTo(2)));
+        assertThat(counts.get(1), is(equalTo(1)));
+        assertThat(counts.get(2), is(equalTo(4)));
+    }
+
+    private AlertNode newAlertNode(
+            int pluginId, int level, String name, String uri, int childCount) {
+        Alert alert = new Alert(pluginId);
+        alert.setUri(uri);
+        AlertNode alertNode = new AlertNode(level, name);
+        alertNode.setUserObject(alert);
+        for (int i = 0; i < childCount; i++) {
+            alertNode.add(new AlertNode(level, name));
+        }
+        return alertNode;
     }
 
     @Test