+++ addOns/reports/CHANGELOG.md
@@ -6,6 +6,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 ## Unreleased
 ### Fixed
  - Address errors when running the OutputSummary job with Automation Framework.
+ - Alert counts to ignore false positives.
 
 ## [0.5.0] - 2021-08-05
 ### Added